criando a tela principal do app
